+++ lib/components/Divider/Divider.treat.ts
@@ -1,5 +1,10 @@
-import { style } from 'sku/treat';
+import { style, styleMap } from 'sku/treat';
 
-export const divider = style({
-  height: '1px',
-});
+export const base = style(theme => ({
+  height: theme.border.width.standard,
+}));
+
+export const weight = styleMap(theme => ({
+  regular: { background: theme.border.color.standard },
+  strong: { background: theme.color.foreground.neutral },
+}));