+++ lib/components/OverflowMenu/OverflowMenu.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useStyles } from 'sku/treat';
+import { useStyles } from 'sku/react-treat';
 import { MenuRenderer, MenuRendererProps } from '../MenuRenderer/MenuRenderer';
 import { OverflowButton } from '../iconButtons/OverflowButton/OverflowButton';
 import { Box } from '../Box/Box';