+++ lib/components/useToast/Toast.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback } from 'react';
 
 import * as styleRefs from './Toast.treat';
-import { useStyles, TreatProvider } from 'sku/treat';
+import { useStyles, TreatProvider } from 'sku/react-treat';
 
 import {
   Stack,