+++ site/src/App/routes/foundations/iconography/iconography.tsx
@@ -141,7 +141,7 @@ const Iconography = () => {
           ) : null}
         </Stack>
 
-        <Inline space={['none', 'medium']}>
+        <Inline space={['none', 'small']}>
           {iconList.map(icon => (
             <IconTile
               key={icon.name}