+++ site/src/App/useScrollLock/useScrollLock.treat.ts
@@ -1,0 +1,5 @@
+import { style } from 'sku/treat';
+
+export const scrollLock = style({
+  overflow: 'hidden',
+});